requirements.txt
nltk==3.7
sklearn==0.0
matplotlib==3.5.2
pandas==1.5
transformers==4.28.1
pytorch-nlp==0.5.0
pytorch-pretrained-bert
gst_calculation==0.1.2
streamlit==1.27.0